Version 01
This is the first version of the page: only one resolution (no RWD yet), created on my own Gulp's configuration, unverified WCAG, counting button clicks does not yet have the function of remembering the number of clicks regardless of refreshing the page, there is also no possibility of resetting the number of clicks (temporarily after 5 clicks, the number of clicks is reset automatically). The first version still requires verification of all points from the task. There are still come comments in code. It's not for production yet.